Implements test for the ProbSum UDAF

+++ src/main/scala/com/doubtless/spark/ProbSumUDAF.scala
@@ -6,27 +6,27 @@ import org.apache.spark.sql.catalyst.encoders.ExpressionEncoder
 import org.apache.spark.sql.Encoder
 
 object ProbSumUDAF
-    extends Aggregator[(Int, BDD), Map[Int, BDD], Map[Int, BDD]] {
-  def zero: Map[Int, BDD] = Map[Int, BDD]((0 -> BDD.True))
+    extends Aggregator[(Double, BDD), Map[Double, BDD], Map[Double, BDD]] {
+  def zero: Map[Double, BDD] = Map[Double, BDD]((0.0 -> BDD.True))
 
   // TODO maybe add the (& ~a._2) clause to the getOrElse part
-  override def reduce(b: Map[Int, BDD], a: (Int, BDD)): Map[Int, BDD] = b.map({
+  override def reduce(b: Map[Double, BDD], a: (Double, BDD)): Map[Double, BDD] = b.map({
     case (sum, bdd) => (sum -> (bdd & ~a._2))
   }) ++ b.map({ case (sum, bdd) =>
     ((sum + a._1) -> ((bdd & a._2) | ((b getOrElse (sum + a._1, BDD.False)) & ~a._2)))
   })
 
-  override def merge(b1: Map[Int, BDD], b2: Map[Int, BDD]): Map[Int, BDD] =
+  override def merge(b1: Map[Double, BDD], b2: Map[Double, BDD]): Map[Double, BDD] =
     b1.toSeq
       .flatMap(tup1 => b2.toSeq.map(tup2 => (tup1, tup2)))
-      .foldLeft(Map[Int, BDD]())((acc, tups) => {
+      .foldLeft(Map[Double, BDD]())((acc, tups) => {
         acc + ((tups._1._1 + tups._2._1) -> ((tups._1._2 & tups._2._2) | (acc getOrElse (tups._1._1 + tups._2._1, BDD.False))))
       })
 
-  override def finish(reduction: Map[Int, BDD]): Map[Int, BDD] =
+  override def finish(reduction: Map[Double, BDD]): Map[Double, BDD] =
     reduction.filter({ case (_, bdd) => bdd != BDD.False })
 
-  def bufferEncoder: Encoder[Map[Int, BDD]] = ExpressionEncoder()
+  def bufferEncoder: Encoder[Map[Double, BDD]] = ExpressionEncoder()
 
-  def outputEncoder: Encoder[Map[Int, BDD]] = ExpressionEncoder()
+  def outputEncoder: Encoder[Map[Double, BDD]] = ExpressionEncoder()
 }
